requirements.txt
aiofiles==0.8.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.3
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==4.0.11
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.1
requests==2.32.3
six==1.10.0
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.8.0
yarl==1.9.4
